pyproject.toml
[project]
name = "modal-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "modal>=1.0.1",
    "open-clip-torch>=2.32.0",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "torch>=2.7.0",
]
